+++ services/api/src/services/storage.py
@@ -1,4 +1,4 @@
-from typing import Any, Optional, Protocol, cast
+from typing import Any, Protocol, cast
 
 from elasticsearch import AsyncElasticsearch, NotFoundError
 
@@ -49,33 +49,3 @@ class BaseElasticsearchService(StorageServiceProtocol):
 
         response = await self.elastic.search(index=resource, body=search_query)
         return [hit["_source"] for hit in response["hits"]["hits"]]
-
-
-class FilmElasticSearchService(BaseElasticsearchService):
-    async def search(  # type: ignore[override]
-        self,
-        resource: str,
-        query: dict[str, Any],
-        page_size: int = 50,
-        page_number: int = 1,
-        sort: Optional[str] = None,
-        **kwargs: Any,
-    ) -> list[dict]:
-        body: dict[str, Any] = {
-            "query": query,
-        }
-
-        if page_size:
-            body["size"] = page_size
-
-        if page_number and page_size:
-            body["from"] = (page_number - 1) * page_size
-
-        if sort is not None:
-            sort_field = sort.lstrip("-")
-            order = "desc" if sort.startswith("-") else "asc"
-            body["sort"] = [{sort_field: {"order": order}}]
-
-        response = await self.elastic.search(index=resource, body=body)
-
-        return response["hits"]["hits"]
